node: extend semantics of wait flag to initial wait
If you have a retrieval and a new prepare back-to-back, it might very
well be that you would send the prepare too early. This is because the
machine info is sent only every two seconds. This is even worsened by
the fact that there is 5sec WAIT-IDLE state after retrieval. This is
there to allow for putting an object back in the real world if it was
removed errorneously, e.g., by another team's robot.

Therefore, to remedy this issue, also wait for up to 10 seconds on the
initial IDLE state of the machine before sending the prepare message. If
the machine is not IDLE after this time report an error.